requirements.txt
psycopg2-binary==2.9.5
PyQt6==6.4.0
PyQt6-Qt6==6.4.0
PyQt6-sip==13.4.0
requests==2.28.1
